Uglify objects and primitives
